+++ src/cli.rs
@@ -1,6 +1,7 @@
 use std::path::PathBuf;
 
 use clap::{ArgEnum, Parser};
+use regex::Regex;
 
 use crate::types::Extension;
 
@@ -16,11 +17,16 @@ macro_rules! arg_env {
     };
 }
 
+const DEFAULT_CLIP_REGEX: &str = r#"(?P<time>[0-9]+(:[0-9]+)+) *.? +(?P<title>.+)"#;
+
+/// Wrapper-tool around `youtube-dl` to create an audio library out of web videos.
+/// Download, clip, and normalize audio streams.
 #[derive(Parser, Debug)]
 pub struct Args {
     /// The playlist ID of all videos to download
-    #[clap(long, env=arg_env!("PLAYLIST"))]
-    pub playlist: String,
+    /// or the ID of the single video to download
+    #[clap(env=arg_env!("ID"))]
+    pub id: String,
 
     /// The path to the output directory
     #[clap(long, env=arg_env!("OUT"))]
@@ -30,16 +36,21 @@ pub struct Args {
     #[clap(long, env=arg_env!("CACHE"))]
     pub cache: PathBuf,
 
-    /// Path to the temporary file. If already existant, content will be lost.
-    /// It must have a verified valid extension (same possible values as --ext)
-    #[clap(long, env=arg_env!("TMP"))]
-    pub tmp: PathBuf,
-
     /// Either keep the entire video or create clips based on timestamps in the description
     #[clap(long, arg_enum, env=arg_env!("SPLIT"))]
     pub split: Split,
 
     /// The file extension to use for the output files. Defines the file container format to use
-    #[clap(long, arg_enum, env=arg_env!("EXT"))]
+    #[clap(long, arg_enum, default_value_t=Extension::Ogg, env=arg_env!("EXT"))]
     pub ext: Extension,
+
+    /// The regular expression for extracting clip timestamps from the description.
+    /// The default value should be able to detect and parse most timestamps.
+    ///
+    /// Must have two named captured groups: `time` and `title`,
+    /// corresponding to the starting timestamp and the title of the clip.
+    ///
+    /// Must use the [Regex crate syntax](https://docs.rs/regex/latest/regex/#syntax)
+    #[clap(long, default_value=DEFAULT_CLIP_REGEX, env=arg_env!("CLIP_REGEX"))]
+    pub clip_regex: Regex,
 }