+++ src/io.rs
@@ -3,7 +3,10 @@ use std::{
     path::{Path, PathBuf},
 };
 
-use anyhow::Result;
+use anyhow::{Context, Result};
+use tempfile::NamedTempFile;
+
+use crate::types::Extension;
 
 pub fn touch<P: AsRef<Path>>(path: P) -> Result<()> {
     OpenOptions::new().create(true).append(true).open(path)?;
@@ -44,4 +47,15 @@ pub fn build_output_path<P: AsRef<Path>>(
     Err(anyhow::anyhow!(
         "Code is broken or you have really REALLY too much files with the same title"
     ))
+}
+
+/// Create a named temporary file and return its handle.
+///
+/// The file destructor will be called at the handle drop.
+/// **As such, one must not simply get the file path and drop the handle.**
+pub fn named_tempfile(extension: Extension) -> Result<NamedTempFile> {
+    tempfile::Builder::new()
+        .suffix(extension.with_dot())
+        .tempfile()
+        .context("Could not create temporary file")
 }